refactor: Split datetime.rs into modular structure
- Created 3 logical modules by functional domain:
  * current.rs - CURRENT_DATE, CURRENT_TIME, CURRENT_TIMESTAMP (54 lines)
  * extract.rs - YEAR, MONTH, DAY, HOUR, MINUTE, SECOND (234 lines)
  * arithmetic.rs - DATEDIFF, DATE_ADD, DATE_SUB, AGE, EXTRACT with helpers (326 lines)
- Converted datetime.rs to mod.rs with re-exports (27 lines)
- All tests passing (700+ workspace tests)
- No functional changes, pure refactoring

Reduces datetime.rs from 690 lines to ~27 lines mod.rs + 3 focused modules
